refactor: replace sidebars to feature/sidebars folder
