+++ frontend/src/app/features/sidebars/edit-sidebar/edit-sidebar.html
