+++ frontend/src/app/features/sidebars/widget-sidebar/widget-sidebar.css
