+++ frontend/src/app/features/sidebars/widget-sidebar/widget-sidebar.ts
@@ -1,10 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { LucideAngularModule } from 'lucide-angular';
-import { WidgetSideBarBtn } from '../../interfaces/widget-sb-btn.interface';
-import { widgetSbBtnsConfig } from '../../configs/widget-sb-btn.config';
-import { Zoom } from '../zoom/zoom';
-import { WidgetService } from '../../services/widget.service';
+import { WidgetSideBarBtn } from '../../../interfaces/widget-sb-btn.interface';
+import { widgetSbBtnsConfig } from '../../../configs/widget-sb-btn.config';
+import { Zoom } from '../../zoom/zoom';
+import { WidgetService } from '../../../services/widget.service';
 
 @Component({
   selector: 'widget-sidebar',
